BaseRawReader now a real base class; binary files now read by BinaryRawReader
